split CommandStream.h and add systrace option
systrace can now be enabled in the CommandStream's Dispatcher, which
allows it to work with all backends.
