+++ filament/CMakeLists.txt
@@ -127,6 +127,7 @@ set(PRIVATE_HDRS
         src/driver/CircularBuffer.h
         src/driver/CommandBufferQueue.h
         src/driver/CommandStream.h
+        src/driver/CommandStreamDispatcher.h
         src/driver/Driver.h
         src/driver/DriverAPI.inc
         src/driver/DriverApi.h
